Cache Inode Rewrite
Rewrite the cache_inode layer extensively.  This rewrite accomplishes
several goals:

  * Replace the call to cache_inode_renew_entry on every entry with a
    single heavy-weight validation on cache_inode_get, and 'trust'
    bits that can be cleared by upcalls from the FSAL.

  * Replace the LRU/garbage collection functions with a new two-level,
    multilane LRU that doesn't impose so much performance overhead.

  * Allow greater concurrency through finer grained locking and
    eliminating the write-lock-then-downgrade-on-every-call pattern.

  * Introduce a weakref scheme, to allow references to entries
    that do not prevent recycling or freeing (such as in directories.)

  * Remove the parent list, vastly simplifying validation.

  * Create a more capable and flexible file descriptor cache and
    method of reaping FDs.

  * Implement a callback function in readdir so as to avoid allocating
    structures and passing them up or requiring the caller to free the
    lock on the directory explicitly.

In order to make rebasing and merging this thing less of an exercise
in tribulation and get rid of double-commits, I have squeezed all the
patches together.  Therefore, here are the contributors:

Matt Benjamin <matt@linuxbox.com> who originated the two-level LRU
design and wrote the initial implementation, hit upon the idea of
using weakrefs and implemented support for them, and doing the
directory integration.

Eitan Ben-Amos <eitanb@il.ibm.com> who provided sharp-eyed code
review and feedback.

Max Matveev <makc@gmx.co.uk> who provided very helpful feedback and
fixed a quite embarrassing bug in readdir.

Frank S. Filz <ffilzlnx@us.ibm.com> who took the time to integrate the
LRU pinning code into the State Abstraction Layer, debugged locking
functionality, and provided feedback in fixing lock ordering in
cache_inode_lru

Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com> and Rong Zeng
<rongzeng@us.ibm.com> who did extensive testing, audited lock use
throughout, and provided a largep ortion of the bug fixes.

+++ src/FSAL/FSAL_GPFS/fsal_convert.c
@@ -512,7 +512,8 @@ fsal_status_t gpfsfsal_xstat_2_fsal_attributes(gpfsfsal_xstat_t *p_buffxstat,
     if(FSAL_TEST_MASK(p_fsalattr_out->asked_attributes, FSAL_ATTR_SIZE))
         {
             p_fsalattr_out->filesize = p_buffstat->st_size;
-            LogFullDebug(COMPONENT_FSAL, "filesize = %llu", p_fsalattr_out->filesize);
+            LogFullDebug(COMPONENT_FSAL, "filesize = %llu",
+                         (unsigned long long)p_fsalattr_out->filesize);
         }
     if(FSAL_TEST_MASK(p_fsalattr_out->asked_attributes, FSAL_ATTR_FSID))
         {
@@ -593,7 +594,8 @@ fsal_status_t gpfsfsal_xstat_2_fsal_attributes(gpfsfsal_xstat_t *p_buffxstat,
     if(FSAL_TEST_MASK(p_fsalattr_out->asked_attributes, FSAL_ATTR_SPACEUSED))
         {
             p_fsalattr_out->spaceused = p_buffstat->st_blocks * S_BLKSIZE;
-            LogFullDebug(COMPONENT_FSAL, "spaceused = %llu", p_fsalattr_out->spaceused);
+            LogFullDebug(COMPONENT_FSAL, "spaceused = %llu",
+                         (unsigned long long)p_fsalattr_out->spaceused);
         }
 
     if(FSAL_TEST_MASK(p_fsalattr_out->asked_attributes, FSAL_ATTR_RAWDEV))
